requirements.txt
Flask==3.1.0
pandas==2.2.3
PyMySQL==1.1.1
python-dotenv==1.1.1
gunicorn==23.0.0
Werkzeug==3.1.3
